reduce PMD warnings down to 2 issues
