+++ cpo-jdbc/src/test/java/org/synchronoss/cpo/jdbc/jta/JdbcXaResourceTest.java
@@ -29,7 +29,7 @@ import org.synchronoss.cpo.helper.ExceptionHelper;
 import org.synchronoss.cpo.jdbc.JdbcStatics;
 import org.synchronoss.cpo.jdbc.ValueObject;
 import org.synchronoss.cpo.jdbc.ValueObjectFactory;
-import org.synchronoss.cpo.jta.CpoXaResource;
+import org.synchronoss.cpo.jta.CpoXaError;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Parameters;
@@ -274,7 +274,7 @@ public class JdbcXaResourceTest {
           cpoXaAdapter1.recover(XAResource.TMSUCCESS);
           fail("XAException should have been thrown");
         } catch (XAException xae) {
-          assertTrue(xae.getLocalizedMessage().startsWith(CpoXaResource.CpoXaError.XAER_INVAL.toString()));
+          assertTrue(xae.getLocalizedMessage().startsWith(CpoXaError.XAER_INVAL.toString()));
         }
       } catch (Exception e) {
         fail(method + ExceptionHelper.getLocalizedMessage(e));
